fix error

+++ minitorch/tensor_ops.py
@@ -275,8 +275,8 @@ def tensor_map(
     ) -> None:
         # TODO: Implement for Task 2.3.
         # raise NotImplementedError("Need to implement for Task 2.3")
-        out_index = np.zeros(MAX_DIMS, np.int16)
-        in_index = np.zeros(MAX_DIMS, np.int16)
+        out_index = np.zeros(MAX_DIMS, np.int32)
+        in_index = np.zeros(MAX_DIMS, np.int32)
         for i in range(len(out)):
             to_index(i, out_shape, out_index)
             broadcast_index(out_index, out_shape, in_shape, in_index)
